chore: prevent chromatic from breaking ci (#115)
Currently our github checks break because chromatic exits non-zero on
changes, this means that since the changes haven't been reviewed on
chromatic, it breaks the CI and fails the rest of the checks. We don't
want this, we want the checks to continue properly, and master should be
auto-ccepted as that's what the PR process is for.

Add two flags to chromatic yarn script to avoid breaking github checks
and auto-accept changes that make it to master.